1226 complete all Notice API
